+++ types/next-auth.d.ts
@@ -5,7 +5,7 @@ declare module "next-auth" {
   interface Session extends DefaultSession {
     user: {
       id: string;
-      role: UserRole.ADMIN | UserRole.CUSTOMER;
+      role: UserRole;
     } & DefaultSession["user"];
   }
 }
@@ -13,6 +13,6 @@ declare module "next-auth" {
 declare module "next-auth/jwt" {
   interface JWT {
     id: string;
-    role: UserRole.ADMIN | UserRole.CUSTOMER;
+    role: UserRole;
   }
 }